refactor(功能实现): auth.ts 文件迁移
